moving the ciphers to a seperate directory
