+++ modules/dev.py
@@ -20,6 +20,8 @@ from past.utils import old_div
 from .util import *
 from .alias import alias_checker
 
+from .ciphers import *
+
 FIREBASE_DYNAMIC_LINK_API_KEY = "AIzaSyAuVJ0zfUmacDG5Vie4Jl7_ercv6gSwebc"
 GOOGLE_URL_SHORTENER_API_KEY = "AIzaSyCBAXe-kId9UwvOQ7M2cLYR7hyCpvfdr7w"
 domain = "yodacli.page.link"
@@ -317,161 +319,6 @@ def search_file(pattern, infile):
         if match:
             yield line
 
-class AtbashCipher:
-    """
-    A class which implements the Atbash Cipher
-    """
-
-    def __init__(self):
-        """
-        Calculates the key dictionary.
-        """
-        self.alphabet = string.ascii_uppercase
-        reverse_alphabet = self.alphabet[::-1]
-        self.key = {" ": " "}
-
-        for index, char in enumerate(self.alphabet):
-            self.key[char] = reverse_alphabet[index]
-
-    def encrypt(self, message):
-        """
-        Atbash Cipher encryption.
-        """
-        encrypted_message = ""
-        message = message.upper()
-
-        for char in message:
-            if char not in self.alphabet and char != " ":
-                click.echo("Atbash only supports ASCII characters.")
-                return
-
-            encrypted_message += self.key[char]
-        return encrypted_message
-    
-    def decrypt(self, message):
-        """
-        As the encrypt and the decrypt "key" are the 
-        same encrypt is its own inverse function.
-        """
-        return self.encrypt(message)
-
-
-class ROT13Cipher:
-    """
-    A class which implements the ROT13 Cipher.
-    """
-    def __init__(self):
-        """
-        Calculates the encryption and decryption alphabet
-        """
-        clear_text_alphabet = list(string.ascii_uppercase)
-        encryption_alphabet = collections.deque(clear_text_alphabet)
-        encryption_alphabet.rotate(13)
-        encryption_alphabet = list(encryption_alphabet)
-
-
-        self.encryption_dictionary = dict(itertools.izip(encryption_alphabet, clear_text_alphabet))
-        self.decryption_dictionary = dict(itertools.izip(clear_text_alphabet, encryption_alphabet))
-
-    def encrypt(self, message):
-        """
-        ROT13 encryption using the previously generate alphabet.
-        """
-        message = message.upper()
-        encrypted_text = ""
-
-        for char in message:
-            if char not in string.ascii_uppercase and char != " ":
-                click.echo("The ROT13 Cipher only supports ASCII characters")
-                return
-
-            if char == " ":
-                encrypted_text += " "
-            else:
-                encrypted_text += self.encryption_dictionary[char]
-        return encrypted_text
-
-    def decrypt(self, message):
-        """
-        ROT13 encryption using the previously generate alphabet.
-        """
-        message = message.upper()
-        decrypted_text = ""
-
-        for char in message:
-            if char not in string.ascii_uppercase and char != " ":
-                click.echo("The ROT13 Cipher only supports ASCII characters")
-                return
-
-            if char == " ":
-                decrypted_text += " "
-            else:
-                decrypted_text += self.decryption_dictionary[char]
-        return decrypted_text
-
-
-class CaesarCipher:
-    """
-    A class which implements the Caesar Cipher.
-    """
-    def encrypt(self, message):
-        """
-        Asks the user for the alphabet shift and then encrypts the text.
-        """
-        message = message.upper()
-        shift = int(click.prompt("The shift value"))
-        encryption_dictionary, _ = self.generate_alphabets(shift)
-        encrypted_text = ""
-
-        for char in message:
-            if char not in string.ascii_uppercase and char != " ":
-                click.echo("The Caesar Cipher only supports ASCII characters")
-                return
-
-            if char == " ":
-                encrypted_text += " "
-            else:
-                encrypted_text += encryption_dictionary[char]
-        return encrypted_text
-    
-    def decrypt(self, message):
-        """
-        Asks the user for the alphabet shift and then decrypts the text.
-        """
-        message = message.upper()
-
-        shift = int(click.prompt("The shift value"))
-        _, decryption_dictionary = self.generate_alphabets(shift)
-        decrypted_text = ""
-
-        for char in message:
-            if char not in string.ascii_uppercase and char != " ":
-                click.echo("The Caesar Cipher only supports ASCII characters")
-                return
-
-            if char == " ":
-                decrypted_text += " "
-            else:
-                decrypted_text += decryption_dictionary[char]
-        return decrypted_text
-    
-    def generate_alphabets(self, shift):
-        """
-        Generates an en and decryption alphabet for a shift.
-        """
-        clear_text_alphabet = list(string.ascii_uppercase)
-        encryption_alphabet = collections.deque(clear_text_alphabet)
-        encryption_alphabet.rotate(shift)
-        encryption_alphabet = list(encryption_alphabet)
-
-
-        encryption_dictionary = dict(itertools.izip(encryption_alphabet, clear_text_alphabet))
-        decryption_dictionary = dict(itertools.izip(clear_text_alphabet, encryption_alphabet))
-        return encryption_dictionary, decryption_dictionary
-
-    
-
-
 @dev.command()
 @click.pass_context
 @click.argument('mode', nargs=1, required=False, callback=alias_checker)
@@ -481,12 +328,17 @@ def ciphers(ctx, mode):
     """
 
     mode = get_arguments(ctx, 1)
+    if mode is None:
+        click.echo("No mode was passed.(choose encrypt or decrypt")
+        return
+    
     _mode = str(mode).lower()
     
+
     cipher_dict = {
-                    "Atbash": AtbashCipher,
-                    "Caesar": CaesarCipher,
-                    "ROT13": ROT13Cipher
+                    "Atbash": atbash.AtbashCipher,
+                    "Caesar": caesar.CaesarCipher,
+                    "ROT13": rot13.ROT13Cipher
                 }
 
     for index, cipher in enumerate(cipher_dict):
